api/requirements.txt

tensorflow==2.11.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.11.0
matplotlib
numpy
